Allow player to walk behind wall tiles
This commit adjusts the collision polygon and z-index of wall-tiles to
allow the player to walk behind them. This provides for more natural
movement. Also, I tightened up the player's collision polygon while I was
about it.

This required adjusting the FOW and screenText z-indices
so they would be drawn over the wall tiles.